Clean up __init__ entry point
Fix bug with False vs false in Python vs JavaScript.
